Move logic from file upload component to a seperate utility file
